👌♻️ put `concatBuffers` in @datadog/browser-core
