+++ packages/worker/package.json
@@ -7,6 +7,9 @@
   "scripts": {
     "build": "rm -rf bundle && webpack --mode=production && ./scripts/build-worker-string.js"
   },
+  "dependencies": {
+    "@datadog/browser-core": "4.45.0"
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/DataDog/browser-sdk.git",
